refactor: refactored endpoints and documentation based on authority/functionality tech: added endpoints for user
