+++ src/main/java/pl/creative/rental_server/core/features/moderator/place/PlaceController.java
@@ -4,14 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-import pl.creative.rental_server.docs.PlaceApi;
-import pl.creative.rental_server.core.features.moderator.item.dto.GetItemDto;
 import pl.creative.rental_server.core.features.moderator.place.dto.EditPlaceDto;
 import pl.creative.rental_server.core.features.moderator.place.dto.GetPlaceDto;
 import pl.creative.rental_server.core.features.moderator.place.dto.InputPlaceDto;
+import pl.creative.rental_server.docs.moderator.PlaceApi;
 
 import java.net.URI;
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -26,16 +24,6 @@ public class PlaceController implements PlaceApi {
                 .buildAndExpand(savedPlace.getId())
                 .toUri();
         return ResponseEntity.created(savedPlaceUri).body(savedPlace);
-    }
-
-    @Override//FOR USER
-    public ResponseEntity<GetPlaceDto> getPlaceById(String placeId) {
-        return ResponseEntity.ok(placeService.getPlaceById(placeId));
-    }
-
-    @Override//FOR USER
-    public ResponseEntity<List<GetPlaceDto>> getAllPlaces() {
-        return ResponseEntity.ok(placeService.getAllPlaces());
     }
 
     @Override
@@ -54,9 +42,5 @@ public class PlaceController implements PlaceApi {
         return ResponseEntity.created(editedPlaceUri).body(editedPlace);
     }
 
-    @Override//FOR USER
-    public ResponseEntity<List<GetItemDto>> getAllItemsByPlaceId(String placeId) {
-        return ResponseEntity.ok(placeService.getAllItemsByPlaceId(placeId));
-    }
 
 }
