+++ src/main/java/pl/creative/rental_server/core/features/user/statusOfItem/UserStatusOfItemController.java
@@ -1,17 +1,17 @@
-package pl.creative.rental_server.core.features.moderator.statusOfItem;
+package pl.creative.rental_server.core.features.user.statusOfItem;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
-import pl.creative.rental_server.docs.StatusOfItemApi;
+import pl.creative.rental_server.docs.user.UserStatusOfItemApi;
 
 @RestController
 @RequiredArgsConstructor
-public class StatusOfItemController implements StatusOfItemApi {
+public class UserStatusOfItemController implements UserStatusOfItemApi {
     private final StatusOfItemService statusOfItemService;
 
 
-    @Override//FOR USER
+    @Override
     public ResponseEntity<?> getAllStatusOfItem() {
         return ResponseEntity.ok(statusOfItemService.getAllStatusOfItem());
     }
