+++ src/main/java/pl/creative/rental_server/core/features/user/statusOfItem/dto/GetStatusOfItemDto.java
@@ -1,4 +1,4 @@
-package pl.creative.rental_server.core.features.moderator.statusOfItem.dto;
+package pl.creative.rental_server.core.features.user.statusOfItem.dto;
 
 import lombok.Data;
 import pl.creative.rental_server.db.entities.StatusOfItem;
