+++ src/main/java/pl/creative/rental_server/docs/moderator/ItemApi.java
@@ -1,71 +1,49 @@
-package pl.creative.rental_server.docs;
+package pl.creative.rental_server.docs.moderator;
 
 import io.swagger.annotations.ApiOperation;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.creative.rental_server.db.entities.StatusOfItem;
-import pl.creative.rental_server.core.features.moderator.item.history.dto.GetItemHistoryDto;
 import pl.creative.rental_server.core.features.moderator.item.dto.FillItemDto;
 import pl.creative.rental_server.core.features.moderator.item.dto.GetItemDto;
+import pl.creative.rental_server.core.features.moderator.item.history.dto.GetItemHistoryDto;
 
 import javax.validation.Valid;
 import java.util.List;
 
-@RequestMapping("/item")
 public interface ItemApi {
 
     @ApiOperation(value = "Adding new item")
-    @PostMapping
+    @PostMapping("/moderator/item/create")
     ResponseEntity<GetItemDto> addItem(@RequestBody @Valid FillItemDto fillItemDto);
 
-    @ApiOperation(value = "Getting the item by given item id")
-    @GetMapping("/{itemId}")
-    ResponseEntity<GetItemDto> getItemByItemId(@PathVariable String itemId);
-
-    @ApiOperation(value = "Getting all items")
-    @GetMapping()
-    ResponseEntity<List<GetItemDto>> getAllItems();
 
     @ApiOperation(value = "Replacing an item")
-    @PutMapping
+    @PutMapping("/moderator/item/replace")
     ResponseEntity<GetItemDto> replaceItem(@RequestBody @Valid FillItemDto fillItemDto);
 
-    @ApiOperation(value = "Getting all items according to given status of the item")
-    @GetMapping("/moderator/items/getItemsByStatusOfItem")
-    ResponseEntity<List<GetItemDto>> getAllItemsByStatusOfItem(@RequestParam StatusOfItem statusOfItem);
-
-    @ApiOperation(value = "Updating status of the item")
-    @PutMapping("/moderator/{itemId}")
-    ResponseEntity<GetItemDto> updateStatusOfItem(@PathVariable String itemId,
-                                                  @RequestParam StatusOfItem newStatusOfItem,
-                                                  @RequestBody(required = false) String commentToEvent);
-
     @ApiOperation(value = "Deleting the item by given item id")
-    @DeleteMapping("/{itemId}")
+    @DeleteMapping("/moderator/item/delete/{itemId}")
     ResponseEntity<?> deleteItem(@PathVariable String itemId, @RequestBody(required = false) String commentToEvent);
 
     @ApiOperation(value = "Deleting the item by force (if item is borrowed by someone) by given item id")
-    @DeleteMapping("/moderator/{itemId}/forceDeleteBorrowedItem")
+    @DeleteMapping("/moderator/item/{itemId}/forceDeleteBorrowedItem")
     ResponseEntity<?> forceDeleteBorrowedItem(@PathVariable String itemId,
                                               @RequestBody(required = false) String commentToEvent,
                                               @RequestParam(required = false, defaultValue = "false") Boolean borrowedItemForceDeleting);
 
-    @ApiOperation(value = "Getting all borrowed items")
-    @GetMapping("/moderator/items/borrowed")
-    ResponseEntity<List<GetItemDto>> getBorrowedItems(@RequestParam boolean borrowed);
 
     @ApiOperation(value = "Updating a place of the item")
-    @PatchMapping("/moderator/{itemId}/{newPlaceId}/changePlaceOfItem")
+    @PatchMapping("/moderator/item/{itemId}/{newPlaceId}/changePlaceOfItem")
     ResponseEntity<GetItemDto> updatePlaceOfItem(@PathVariable String itemId, @PathVariable String newPlaceId,
                                                  @RequestBody(required = false) String commentToEvent);
 
     @ApiOperation(value = "Updating a borrower of the item")
-    @PatchMapping("/moderator/{itemId}/{newAccountId}/changeBorrowerOfItem")
+    @PatchMapping("/moderator/item/{itemId}/{newAccountId}/changeBorrowerOfItem")
     ResponseEntity<GetItemDto> updateBorrowerOfItem(@PathVariable String itemId, @PathVariable Long newAccountId,
                                                     @RequestBody(required = false) String commentToEvent);
 
     @ApiOperation(value = "Getting all history of the item by given item id")
-    @GetMapping("/moderator/{itemId}/getHistoryOfItem")
+    @GetMapping("/moderator/item/{itemId}/getHistoryOfItem")
     ResponseEntity<List<GetItemHistoryDto>> getAllHistoryOfItem(@PathVariable String itemId);
 
 }
