+++ src/main/java/pl/creative/rental_server/docs/moderator/PlaceApi.java
@@ -1,29 +1,19 @@
-package pl.creative.rental_server.docs;
+package pl.creative.rental_server.docs.moderator;
 
 import io.swagger.annotations.ApiOperation;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.creative.rental_server.core.features.moderator.item.dto.GetItemDto;
 import pl.creative.rental_server.core.features.moderator.place.dto.EditPlaceDto;
 import pl.creative.rental_server.core.features.moderator.place.dto.GetPlaceDto;
 import pl.creative.rental_server.core.features.moderator.place.dto.InputPlaceDto;
 
 import javax.validation.Valid;
-import java.util.List;
 
 public interface PlaceApi {
 
     @ApiOperation(value = "Adding new place")
     @PostMapping("/moderator/place/create")
     ResponseEntity<GetPlaceDto> addPlace(@RequestBody @Valid InputPlaceDto inputPlaceDto);
-
-    @ApiOperation(value = "Getting the place by given place id")
-    @GetMapping("/place/{placeId}")
-    ResponseEntity<GetPlaceDto> getPlaceById(@PathVariable String placeId);
-
-    @ApiOperation(value = "Getting all places")
-    @GetMapping("/moderator/places")
-    ResponseEntity<List<GetPlaceDto>> getAllPlaces();
 
     @ApiOperation(value = "Deleting the place by given place id")
     @DeleteMapping("/moderator/{placeId}/delete")
@@ -33,8 +23,5 @@ public interface PlaceApi {
     @PutMapping("/moderator/{placeId}/update")
     ResponseEntity<EditPlaceDto> updatePlace(@RequestBody @Valid InputPlaceDto inputPlaceDto, @PathVariable String placeId);
 
-    @ApiOperation(value = "Getting all items by given place id")
-    @GetMapping("/moderator/{placeId}/getItemsByPlaceId")
-    ResponseEntity<List<GetItemDto>> getAllItemsByPlaceId(@PathVariable String placeId);
 
 }
