+++ src/main/java/pl/creative/rental_server/docs/user/UserStatusOfItemApi.java
@@ -1,10 +1,10 @@
-package pl.creative.rental_server.docs;
+package pl.creative.rental_server.docs.user;
 
 import io.swagger.annotations.ApiOperation;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 
-public interface StatusOfItemApi {
+public interface UserStatusOfItemApi {
 
     @ApiOperation("Getting all status of item")
     @GetMapping("/getAllStatusOfItem")
